Extend diag of CAN driver

+++ Drivers/sx1262/sx1262_commands.c
@@ -673,7 +673,7 @@ bool sx1262_test_command(int32_t argc, char* argv[]) {
         Sx1262Instance.tx_done_cnt = 0;
         Sx1262Instance.rx_done_cnt = 0;
         for(try_num = 1; try_num <= try_cnt; try_num++) {
-            snprintf(tx_array, sizeof(tx_array), "ping_%u", try_num);
+            snprintf(tx_array, sizeof(tx_array), "ping_%u", (uint32_t)try_num);
             tx_array_len = (uint16_t)strlen((char*)tx_array) + 1U;
             LOG_INFO(LORA, "send %u/%u [%s] pause %u ms", try_num, try_cnt, tx_array, wait_pause_ms);
             res = sx1262_start_tx(tx_array, tx_array_len, 0);
@@ -688,7 +688,6 @@ bool sx1262_test_command(int32_t argc, char* argv[]) {
     }
     return res;
 }
-
 
 bool sx1262_statistic_command(int32_t argc, char* argv[]) {
     bool res = false;
